fix build: 	install runtime lib in correct location. 	fix one bad type definition in defs.h. 	clear out $GOROOT/lib in clean.bash.
TBR=r
OCL=26691
CL=26691

+++ src/runtime/darwin/amd64/defs.h
@@ -80,9 +80,9 @@ typedef struct MachPort MachPort;
 struct MachPort {
 	uint32 name;
 	uint32 pad1;
-	uint32 pad2;
-	uint32 disposition;
-	uint32 type;
+	uint16 pad2;
+	uint8 disposition;
+	uint8 type;
 };
 
 typedef struct StackT StackT;
